requirements.txt
torch
numpy
opencv-python
matplotlib
torchvision
tqdm
Pillow
pandas
scikit-learn
Flask
tensorboard
easydict
onnx
onnxruntime
